Add uniform crossover

+++ cpu-naive/src/adjuster.rs
@@ -44,9 +44,9 @@ impl Adjuster {
         Self { max_no_improvement }
     }
 
-    pub fn adjust(&self, stats: &GenerationStats, mutation: &mut f32, crossover: &mut f32) {
-        let n = stats.no_improvement as f32;
-        let m = self.max_no_improvement as f32;
+    pub fn adjust(&self, stats: &GenerationStats, mutation: &mut f64, crossover: &mut f64) {
+        let n = stats.no_improvement as f64;
+        let m = self.max_no_improvement as f64;
         let r = 1.0 + (n - m) / m * 0.001;
         eprintln!("r: {}", r);
